chore: refactor render-rules tool
